SellerForm

+++ src/model/dao/SellerDao.java
@@ -1,16 +1,14 @@
 package model.dao;
 
 import java.util.List;
-
 import model.entities.Department;
 import model.entities.Seller;
 
 public interface SellerDao {
-
 	void insert(Seller obj);
 	void update(Seller obj);
-	void deleteById(Integer id);
-	Seller findById(Integer id);
+	void deleteById(Long id);
+	Seller findById(Long id);
 	List<Seller> findAll();
 	List<Seller> findByDepartment(Department department);
 }
